requirements.txt
cryptography==36.0.2
fastapi==0.75.1
gunicorn==20.1.0
httpx==0.22.0
pydantic==1.9.0
python-dotenv==0.20.0
uvicorn[standard]==0.17.5
xlwings==0.27.6

# Only required for Google Directory
google-api-python-client==2.42.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
